Fix macOS clang build errors across lowerer and tests
Add [[maybe_unused]] to reference members that Apple clang flags

under -Werror -Wunused-private-field in three IR lowerer headers.

Remove genuinely unused private fields from StatementsExpr and

SumHelpers contexts.

Add missing <algorithm> and <string> includes and resolve

std::to_string ambiguity in test files.

+++ tests/unit/compile_run/bindings/test_compile_run_bindings_and_examples.cpp
@@ -13,7 +13,7 @@ static std::string spinningCubeBackendProbeCacheSignature(const std::filesystem:
     ec.clear();
     const auto lastWrite = std::filesystem::last_write_time(path, ec);
     signature += "\nmtime=";
-    signature += ec ? "missing" : std::to_string(lastWrite.time_since_epoch().count());
+    signature += ec ? "missing" : std::to_string(static_cast<long long>(lastWrite.time_since_epoch().count()));
   };
 
   appendPathStamp(cubePath);
